SpaceOrganizer/backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
google-cloud-vision==2.4.2
opencv-python-headless==4.5.5.62
scipy==1.7.3
ortools==9.1.9490
Plotly==5.5.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
gunicorn==20.1.0
pytest==7.0.1
python-dotenv==0.19.2
tensorflow==2.8.0
torch==1.10.2
Pillow==9.0.1
SQLAlchemy==1.4.27
Flask-Migrate==3.1.0
Flask-Testing==0.8.1
Flask-Script==2.0.6
python-jose==3.3.0
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
MarkupSafe==2.0.1
numpy==1.22.3
pandas==1.4.1
requests==2.27.1
six==1.16.0